Added RMI implementation
- Introduced `Client` interface which mimics `KademliaRPC` interface without the `Host` requirement in the API to serve as the foundation for actual implementors
- Introduced `RemoteClient` interface which wraps `Client` and `Remote` interfaces together and acts as the building block for remote objects to interact with each other via RMI
- Introduced `RemoteClientImpl`, the main implementation of `RemoteClient`. It simply delegates actual implementation to an injected handler, which is just another `Client`. Basically, there should be 1 `RemoteClient` per machine, and it's handler is currently an instance of `KademliaClient`
- Updated `KademliaClient` to contain a remote object so that it can receive RMI from other machines. The remote object is an instance of `RemoteClientImpl` at this time.
- Introduced `KademliaRPCImpl`,  which conveniently wraps RMI framework logic
- Updated all references to `key` from `int` to `long`, as this usually refers to a node's IP address which could be 64 bits.

This commit contains a long list of changes, but I didn't want to waste time splitting it up into more logical commits. The main takeaway is that the native Java RMI libraries can help achieve remote invocations of our program across different computers. Usually RMI involves a client and a server. In this instance, each instance of our program is acting as a client and a server (for the purposes of the Java RMI framework) in order to send and receive messages